Fix call to htcondor python binding
With this change the cscollector now properly executes send_command with a peaceful condor off.
There were some issues testing on htcs-s1 as the jobs from the queue never actually claimed a
machine. I will test it further on htcs-master as I am not sure the machines are properly
executing the peaceful shutdown.