update print string for job class and refine logic for manager class
The JobManager class now maintains 4 dictionaries, one for scheduled jobs, one of unscheduled ones and
two more that index the same lists by user.

Logic for each stub function has been implemented, the only one that i know needs work is the update_jobs function.
We still need to decide how we want to handle the periodic dump of jobs. When an instance of the job class is
instantiated it is given the default unscheduled state (regardless if it is already in one of the dicts).

Perhaps the solution is as simple as checking the currently maintained dictionaries for the jobID to see which
dict it would be appropriate to add the given job to.