revamp second layer of django authentication
It is not possible in django to append columns to the auth_user table and make them available
under the 'User' model. This means that we will have to match the authorized user from the
metadata each time a connection is made to either csv2_user.username or csv2_user.cert_dn.