add group name logic to csjobs
csjobs now inspects the job classad to make sure there is a specified group name.
If no group name is specified, it tried to automatically resolve it by assigning
the group name associated with the submitting user. If the user is registered to
more than one group it gives up and ignores the ad after a log message.

When I was testing this code in the command line i was unable to update the ads
running as my personal user, but had no issues running as the root user.

I suspect we will have to give some sort of elevated condor permission to the
user this service runs under or it will be unable to edit the class ads to add
the group name.