remove unused proccess functions, update config to use config stored in database
This update also removes the cleanup process and instead implements a process for each type of metadata
The configuration files have been updated to pull in the configuration options from the database.

These changes in general also mean that there is only one sleep cycle per type of metadata.
The VM polling routine is unchanged and is still two proccesses for polling and cleanup.