remove redirects
Redirects in django do not allow you to pass parameters outside ones placed in the url.
If we are to make the change away from calling other views directly in a respnonce
to ensure the url is accurate it will take soem careful planning and will also need
a well defined url structure.