add support for no-group and non-csv2 jobs in job poller
I've added an option in the yaml config for csv2-jobs that allows the user to select a default group to
assign to a job. This should help facilitate users that already had a condor batch system running and
want cloud scheduler to run (or not run) some of the existing jobs.