add glintv2 support for csv2
This commit should properly serve the glint web pages with a running set of celery workers
The glint static files still need to be properly setup as well as the cache/working directory.
Ideally the latter will be initially setup by the ansible deployment.

There can be some issues getting the database in a proper state due to the table already existing
The dirty solution is to delete all the tables used by glint and have the manage.py create them again
so that it is in the proper state since some of the tables are created just for glint. However,
since we are constantly backingup the schema we should now have all the tables in this database
and we can run a fake django migration and everything should work if the tables for glint are present.