add static variables for group and cloud name for error reporting
When encountering a sqlalchemy error and attempting to handle it can cause nested errors if you attempt
to access one of the base object's variables. To get around this and avoid thread crashes i've assigned
the group and cloud name at the start of each loop to use in the error messages.