fix performance issue in job poller
This took a very long time to track down but the issue ended up being that the attribute list was growing.
The list was passed into the 'trim_keys' function where group_name was appended so it wouldn't be trimmed.
However since this was teh original attribute list and not a copy- every single job added on 'group_name'
to the attribute list and thus condor was trying to project over group_name thousands of times. With this
small change the job poller cycle has returned to a matter of seconds.