update status poller to use new configuration method
The status poller still creates it's own db objects and should still be updated to use the
ones defined in the new configuration.
This will be more complicated for the other pollers which have multiple threads as there
is some shared configuration that the main function needs to setup logging and intervals
but each thread has it's own configuration and connection needs so it may still be better
to define a database connection in each thread and leave the config connection for the main thread.