add general instance type managment to ec2FlavorPoller
With this commit the flavor poller now checks on the local json files and makes sure they have been refreshed within the last 7 days.
If they have not a new version is downloaded then parsed into the ec2_instance_types table that will server as a springboard
to populate cloud_flavors when an amazon cloud has been added.