update ec2 poller and ec2cloud to properly boot and track ec2 vms
With this commit cloudscheduler will boot vms on amazon and the vms will register with condor
but the machine poller still needs an update to properly track slots from ec2 vms.

A classad attriibute 'csv2_id' will be added to simplify the classifications without having to use
the VM's hostname.