update glint to use new database centric transaction method
This commit has a very elementry images page with buttons to transfer and delete.
The buttons succuessfully queue transfer requests and the delete buttons are fully functional.
A confirmation alert should be added for delete and a warning for when you try to delete
a shared image because most of the time you will be unable to delete them.

The celery tasks have not yet been asynchronously launched but the database rows representing
the transactions are properly queue'd and now just need to be hooked up to the celery tasks.

I've changed the configuration in settings.py and in the database for celery to properly use
amqp as its queuing system- however, it doesn't yet use any form of authentication.

The last steps are to hook up the celery tasks, fix the css of the page and add warnings.
It would also be nice to add a search like the old version had, and maybe an option to hide
public/shared images.