update apel accounting table and ec2 state references to fix #264
These updates address the issue of spot instances having an empty hostname which caused the trigger inserting into the apel accounting table to fail.
To fix this I've changed the 3rd part of the primary key to use VMID instead of hostname since spot instances will always have a value there.

The second change is in responce to what seems like a slightly different data structure in the responce from amazon to get a VMs state.

These changes have passed preliminary tests but should be subject to extensive testing to distill out any other issues.

This also may have an effect on the apel accounting values for spot instance VMs since the hostname will likely never be added and it may be somethign we want to revisit later.