update foreign job counts in csjobs
I've updated the row in csv2_job_schedulers to insert the total foreign jobs instead of the 'held' jobs which are the jobs taht have a valid group name but violate one of the other rules such as being a valid user.

With this change we should consider holding bad jobs again in the future.