upgrade condor poller na to use sqlalchemy free version of the database config
other supporting files had to be upgraded to use the same version of the db_config.

These chagnes should eventually be propogated across the rest of the csv2 system to remove the sqlalchemt dependencies.
This has essentially created a large amount of technical debt for csv2 that should be allivated.