Add dynamic process type to ProcessMonitor_na.py
This change allows the ProcessMonitor to start and stop new processes dynamically as groups and clouds are added or removed from csv2.
This change is in response to orphaned grandchildren processes spawned to execute condor commands. The orphaned child tasks caused the machine command poller thread to never spawn new children to execute retire tasks.
This change will likely need to be implemented in the alchemy version of ProcessMonitor to support the same terminate tasks from csmain.

These chagnes are untested.