fix keypair transfer bugs
This commit fixes 2 issues.
Firstly, if a key transfer was requested and it tried to get the source key from a cloud that was offline the transfer would fail.
Now it will try all possible keypair sources until one works or they all fail.
Secondly i've fixed the redirect after a transfer was requested to properly return to the same group's matrix