update defaults replication process in openstackpoller
This commit contains several changes to the process.

Firstly it will no longer transfer to//from disabled clouds or clouds that have  no network connection.
Secondly i've added more debug logging to make it easier to understand whats going wrong if there is issues with the process in the future.
Lastly i've seperated keypairs and images into seperate for loops so a failure on one will not distrupt the transfer of the other.