fix image upload from file
The previous changes made to properly specify a list of clouds when uploading an image conflicted with the way request toobelts handles input.
It tries to treat any list dictionary entries as a file to upload and thus was confused by a list of clouds.
The string processing to get the list of clouds has now been moved to the server side circumventing this limitation.