upgrade lowerdash regex filter and fix ec2 filterer crashes
This update contains 2 changes.
Firstly the upgrade to the lowerdash regex that will now properly check for all valid strings that dont start or end with a dash.
Secondly the schema update contains a change to cloud_images that allows null for the created_at field because amazon images lack this metadata.