update job group scheduling in csmain, reduce debug logging in several files, fix kill retire behaviour
This commit contains several updates for 2.8.1:
-csmain did not rotate through the groups of jobs (alias) for each group and thus the first job in the list always got the resouces first.
i've updated this behaviour to use the same method that is used for groups using the while counter as a modulus for the starting point

-several debug loggin statements where still at an elevated log level, some have been updated to be more accurate and changed to a more apprpriate
level

-The removal of sqlalchemy introduced a bug from the python sql library we are now using, when providing db_config with a compound sql statement
we must pass the multi option to the database call. However, this caused problems with temporary tables as the multi option can use different
connections to issue the sql commands making the temp tables unavailable occasionally. I've rewritten the routine to issue only single staments
per database execute resolving the issue.