update csmain vm_slot_auditor process to modify database when an inconsistency is found.
This change has been tested under a low level test job load with manual edits to the database to create
inconsistencies. The only outstanding change would be to make the sleep cycle configurable instead
of a hard coded 30s. In general this thread should not be run often since it compute heavy
and only meant to catch mistakes for retiring vms. 300s would probably be a more appropriate cycle time.