add watchdog process to process monitor
As part of the Processmonitors regular operation a watchdog check has been implemented
that routinely checks on the status of all active subprocesses. If the watchdog determines
any processes have been inactive it will terminate the subprocess and restart them.

Since some subprocess that use the Processmonitor have a very long sleep cycle
or are only brought out of sleep under specific conditions I have added an
exemption field that can be passed to the Processmonitor constructor. The
processes designated in the list will be exempt from watchdog restarts.

Each process registers a heartbeat with the csv2_watchdog table and updates the timestamp
each cycle. If the watchdog sees the timestamp hasn't been updated in the last 60 minutes
the process will be terminated and restarted. This should hopefully help the part per
million errors that sometimes cause the openstack or condor processes to stall without
crash.

The 60 minute window has been statically coded but it may be a good idea to change it to
a configuration value- for now 60m is a good value as it is much longer than and default
sleep cycle configuration.

It may be desirable to further expand the watchdog to collect more data abotu each process
and possibly add an admin-email feature when it detects a stall to alert the sysadmin.