update ec2 components to properly use new db_config return values
this commit also fixes several table name definitions and small typo errors.
The changes to the db_config return values are checked but no exceptions are raised
nor do the functions interrupt the loop or return where appropriate
this shouldnt effect the default behaviour of these components but it does leave
more advanced error handling on the table.