service_enable_condor_poller: accommodate more robust pip syntax parsing.
Almalinux9 has python 3.9 and pip is 24.1.  Compared to earlier versions of
pip (such as pip 21.3.1 on CentOS7), this combination has tightened the
syntax of pip parameters such that a version matching clause within a
version specifier must now include a specific version number.

On earlier systems, specifying a version matching clause of "htcondor==" had
the side effect of printing all available htcondor python module versions as
part of the error message.  service_enable_condor_poller would use this list
to identify the version to install based on the system's htcondor
installation.  Since "htcondor==" is now rejected as a parameter error, the
list of available module versions is not shown.  Consequently
service_enable_condor_poller fails to identify a version of the htcondor
python module to install when running on Almalinux9 and therefore doesn't
install anything.  This in turn prevents condor_poller from starting.

Experimentation showed that "htcondor==0" produces the expected output on
CentOS7 and Almalinux9 systems.  This means there is no need to use
different pip parameter syntax depending on the pip version in use.