condor_poller: don't assume condor_worker_cert is defined.
If condor_worker_cert is not set in the configuration file, the respective
key won't be present in config.condor_poller.  This will cause the
conditional in worker_gsi_poller() to fail because the existing expression
assumes that the key exists in some form.

When there is no worker certificate (as is the case if tokens are being
used), this issue can be worked around by setting condor_worker_cert to the
empty string in the configuration file.  However, this triggers collateral
error messages because the code assumes that if defined, condor_worker_cert
is a valid filename.  Addressing the underlying problem is therefore
worthwhile.